feat: create WeatherIcon component
