+++ src/components/WeatherIcon/index.jsx
@@ -1,0 +1,1 @@
+export { default, validWeather} from './WeatherIcon'
